revert: move revertccl out of ccl
Release note: none.
Epic: none.
