+++ pkg/BUILD.bazel
@@ -75,7 +75,6 @@ ALL_TESTS = [
     "//pkg/ccl/partitionccl:partitionccl_test",
     "//pkg/ccl/pgcryptoccl/pgcryptocipherccl:pgcryptocipherccl_test",
     "//pkg/ccl/pgcryptoccl:pgcryptoccl_test",
-    "//pkg/ccl/revertccl:revertccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
     "//pkg/ccl/serverccl/adminccl:adminccl_test",
     "//pkg/ccl/serverccl/diagnosticsccl:diagnosticsccl_test",
@@ -301,6 +300,7 @@ ALL_TESTS = [
     "//pkg/raft/rafttest:rafttest_test",
     "//pkg/raft/tracker:tracker_test",
     "//pkg/raft:raft_test",
+    "//pkg/revert:revert_test",
     "//pkg/roachpb:roachpb_disallowed_imports_test",
     "//pkg/roachpb:roachpb_test",
     "//pkg/roachprod/cloud:cloud_test",
@@ -951,8 +951,6 @@ GO_TARGETS = [
     "//pkg/ccl/pgcryptoccl/pgcryptocipherccl:pgcryptocipherccl_test",
     "//pkg/ccl/pgcryptoccl:pgcryptoccl",
     "//pkg/ccl/pgcryptoccl:pgcryptoccl_test",
-    "//pkg/ccl/revertccl:revertccl",
-    "//pkg/ccl/revertccl:revertccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
     "//pkg/ccl/securityccl/fipsccl:fipsccl",
@@ -1604,6 +1602,8 @@ GO_TARGETS = [
     "//pkg/release:release",
     "//pkg/repstream/streampb:streampb",
     "//pkg/repstream:repstream",
+    "//pkg/revert:revert",
+    "//pkg/revert:revert_test",
     "//pkg/roachpb:roachpb",
     "//pkg/roachpb:roachpb_test",
     "//pkg/roachprod/cloud:cloud",
